# lib/api-zod/src/generated/types/articleUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Royal Pulse News API
 * OpenAPI spec version: 0.1.0
 */
import type { ArticleUpdateStatus } from './articleUpdateStatus';

export interface ArticleUpdate {
  title?: string;
  content?: string;
  excerpt?: string;
  category?: string;
  author?: string;
  imageUrl?: string;
  status?: ArticleUpdateStatus;
  isFeatured?: boolean;
  isBreaking?: boolean;
  readTime?: number;
}