# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Royal Pulse News API
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  articleId: number;
  authorName: string;
  /** @nullable */
  authorEmail?: string | null;
  content: string;
  createdAt: Date;
  approved: boolean;
}