# lib/api-zod/src/generated/types/tipResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Royal Pulse News API
 * OpenAPI spec version: 0.1.0
 */

export interface TipResult {
  success: boolean;
  message: string;
}